k_closest_points

+++ heap/kth_smallest_sorted_matrix.cc
@@ -12,10 +12,10 @@ public:
     
     int kthSmallest(vector<vector<int>>& matrix, int k) {
         priority_queue<int, vector<int>> pq;
-        for (int i = 0; i < matrix.size(); ++i) {
-            for (int j = 0; j < matrix.size(); ++j) {
+        for (size_t i = 0; i < matrix.size(); ++i) {
+            for (size_t j = 0; j < matrix.size(); ++j) {
                 pq.push(matrix[i][j]);
-                if (pq.size() > k) {
+                if ((int)pq.size() > k) {
                     pq.pop();
                 }
             }
